made all the api functions work

+++ src/APIFunctions/deleteMedia.js
@@ -1,14 +1,25 @@
-const deleteMylistMedia = async (mediaId)  => {
+import { url } from "../constants/constants"
+
+export const deleteMylistMedia = async (mediaId)  => {
 	try {
+		const token = localStorage.getItem('token')
+		if (!token) {
+			throw new Error('JWT token not found in local Storage.')
+		}
 		const response = await fetch(
-			'',
+			url + 'bingebox/mylists',
 			{
 				method: "DELETE",
 				headers: {
 					"Content-Type": "application/json",
+					'Authorization': `Bearer ${token}`,
+					"Access-Control-Allow-Origin": "*",
+					"Access-Control-Allow-Methods": "GET, DELETE, PUT, POST, OPTIONS",
 				},
+				body: JSON.stringify({mediaId: mediaId})
 			}
 		)
+		console.log('mediaId:' , response)
 
 		if (response.ok) {
 			console.log("media deleted successfully")
@@ -21,7 +32,7 @@ const deleteMylistMedia = async (mediaId)  => {
 }
 
 
-const deleteWatchedMedia = async (mediaId) => {
+export const deleteWatchedMedia = async (mediaId) => {
 	try {
 		const response = await fetch(
 			``,
@@ -43,4 +54,3 @@ const deleteWatchedMedia = async (mediaId) => {
 	}
 }
 
-export default {deleteWatchedMedia, deleteMylistMedia }
